modify `make_execute_request` to be async and return the input root's `DirectoryDigest` (#17370)
Modify `make_execute_request` to be async, take a `Store` as a parameter, and to return the input root to use. This is prework for https://github.com/pantsbuild/pants/pull/17290 which needs the ability to make a new input root digest if a wrapper script needs to be added to the input root. (It will also be useful for other wrapper script applications, such as to solve the issue at  https://github.com/pantsbuild/pants/pull/17367#discussion_r1006760027 for remote execution.)