Switch `pytest` plugin to a custom partitioner rule.
The custom partitioner still creates one partition per field-set, but
the runner rule should now be capable of testing an arbitrary number of
input field-sets. As part of this, I've defined the metadata type for
`pytest` partitions and shifted some logic for calculating metadata
fields from the runner rule into the partitioner rule.