Adds warning if lint/fmt/fix runs on a target using a non-local environment. (#17359)
This adds a warning if a target-based `fmt`/`lint`/`fix` operation runs against a target that specifies a non-local environment. 

To property use environments, we'll need a way of making the partitioning rules a bit more pluggable, or decide that individual back-end implementations will be responsible for deciding where they run (which is probably also not great).

Addresses #17129.