Add `batch_compatibility_tag` field to `python_test` for partitioning.
If two `python_test` targets have the same value for
`batch_compatibility_tag`, then they are eligible to be batched into the
same `pytest` process. They might _not_ be batched in the same process
if:

    * There are "too many" compatible tests as determined by the
      `[test].batch_size` config option
    * Compatible tests have some incompatible Pants metadata (i.e.
      different `extra_env_vars` or `xdist_concurrency`)

Tests that don't set the new field are assumed to be incompatible with
all others, and continue to run one-target-per-process.