Fix dirtying of uncacheable nodes (Cherry-pick of #17079) (#17396)
As reported in #17060, the same `@rule` running within a single "session"/"run" of Pants was not correctly re-running due to changed inputs. The `@rule` is uncacheable (via its output type being `EngineAwareReturnType.cacheable = False`), for the purposes of a logging side-effect.

The `@rule` was failing to re-run due to a bug in the dirtying of uncacheable rules that had likely existed since #13199 split `Node::restartable` out from `Node::cacheable`. When the `restartable` property was introduced, it took over controlling whether a Node could be interrupted while running. But the `Entry::dirty` codepath was still incorrectly preventing `uncacheable` `Node`s from being dirtied.

Fixes #17377.