Flip the default for `pull` on `docker_image` to `False` (#17459)
Closes #17458

57746408819a19862123600447ea28cee8021064 added the `pull` field to `docker_image`, with a default of `True`. Changing the default behavior in that way caused multi-stage builds with "transient" base images to break, because Docker can't pull the base images that only exist locally. This commit fixes the problem by flipping the default to `False`.

IMO a default of `False` makes sense regardless of the bug it fixes, because it aligns with the default behavior of `docker build`. Users who want to always pull by default can use `__defaults__` to achieve the same effect.

I've added a simple integration test to guard against regressions here in the future.