Isolate config seed values from the DEFAULT section. (#17474)
Pants config files support interpolation of certain seed values, such as `%(buildroot)s` 
and `%(env.NAME)s` into user-supplied values.  They also support a DEFAULT section, 
that provides values for keys not specified in the relevant section.

Both of these features are standard features of Python's ini file parser, which we
carried over, and implemented ourselves, when we switched from ini to toml for our config.

Unfortunately, these two features are entangled: the ini parser allows interpolations that 
reference values in the same section, or in the DEFAULT section. So we injected the seed 
values into the DEFAULT section to allow them to participate in interpolation. And our toml
reimplementation did the same.

An unintended consequence of this is that if you create an option with the name of one of the
seed values (e.g., `env` or `user`), and don't set it explicitly in its section, you will read the 
seed values from the DEFAULT section. In the case of `env` this default will be all the env vars 
seen by pants!

This change:
- Modifies the implementation of interpolation, so that we don't use the DEFAULT section as the
   source of interpolation values, but a separate data structure consisting of the seed values
   and a copy of the DEFAULT section.
- Improves the config unittests by introducing a `_compare` helper that more rigorously compares
   expected and actual configs (previously we checked that the values we expected were present, 
   but not that all values present were expected).