Subset CTs from partition metadata in `pylint` runner rule.
The coarsened targets passed to the `pylint` "runner" rule will
encapsulate all the dependencies of the original partition, which can be
many more dependencies / files than are needed to lint the specific
subset of the partition passed to the rule. Filter the CTs to avoid
burning time on building too big of a requirements pex / materializing
too many input files.