Disambiguate Python module providers by level of ancestry.
I.e., if we're looking for a provider for foo.bar, and we have
one for foo.bar and one for foo, take the former over the latter.

Previously we would see this as an ambiguity, and not infer a dep.
This happened to a real user, in a situation where a namespace
package was split between first-party and third-party code:
https://github.com/pantsbuild/pants/discussions/17286

This change involves a distinction between ModuleProvider and a
new PossibleModuleProvider class, that also tracks the ancestry.
Therefore {First,Third}PartyPythonModuleMapping now encapsulate
the underlying dict instead of extending the dict. This clarifies
that their role is to use the dict to produce other data, and
they are not intended to be queried like dicts.