Filter coarsened targets to only those that are relevant in `pylint` runner rule (#17487)
Closes #17486

Prior to this commit on the 2.15.x line, calculation of coarsened targets for `pylint` inputs would run as part of the "partitioner" rule. The computed targets were tracked as part of partition metadata and passed to the "runner" rule, where they were used to build the transitive requirements pex / to hydrate sources in the `pylint` sandbox.

This approach did not account for the fact that `lint` partitions get re-batched according to `[lint].batch_size`. Re-batching doesn't modify partition metadata, so we would end up trying to hydrate a full partition's worth of transitive dependencies for each ~small `pylint` batch, instead of only those transitive dependencies relevant to each batch's final elements. In my large monorepo, this consistently caused `pantsd` to freeze and eventually be OOM killed.

Resolving coarsened targets once in the "partitioner" rule is good to avoid other performance issues, so we keep that logic and add a secondary filter in the "runner" rule to select only the CTs that are needed for the batch of files that ultimately get submitted as an input.