Export metadata about a packaged docker image (#17299)
This has the docker target output a `$name.docker-info.json` file when packaged, to end up in `dist/` or for use as a dependency. The file contains the full specification of the packaged image, including each of the `registry/repository:tag` names it was tagged with. This stops users from having to parse the pants output (or query docker itself) if they need this information, such as to place into a cloud deploy template, or to run the image locally.