Mark `pants` as an explicit namespace package (rather than implicit) (#17563)
Now we aren't touching our source tree during tests, which is causing unintended side-effects in https://github.com/pantsbuild/pants/pull/17520, because really we shouldn't be assuming the process sandboxes for tests are mutable (at least in the source tree). Not to mention that lengthy comment in `conftest.py` which should raise alarm.