reproduce ignore rules inconsistency
There's an inconsistency between patterns defined in .gitignore and
patterns defined in pants_ignore for how the filesystem watcher is
invalidating ignored paths.  For paths only in .gitignore, if you
`touch` that node, and the parent directory is not ignored, it still
invalidates ongoing build graph steps.  However, if this path is in
pants_ignore, it is ignored properly.

To reproduce on this branch, comment out the `#/ignored_dir` line
from pants.toml and run the following command:

```
./pants test \
  --extra-env-vars=IGNORED_DIR="$(pwd)"/ignored_dir \
  code_using_ignored_dir/test_writing_files.py
```

You'll see the test never complete because the `touch` inside the test
invalidates the graph.  Comment the line in pants.toml again, and re-run
the command, and you'll see the test run fine.

This is just a demo PR, it is not intended to be merged ever

[ci skip-rust]