shell: add `experimental_test_shell_command` for arbitrary shell-driven tests (#17640)
Add an `experimental_test_shell_command` target type to enable defining arbitrary tests using shell code.

Note: This target is not a unit test of shell sources like `shunit2_tests`; rather, it is a test defined in shell which could test anything which can be packaged and consumed by the other shell targets (e.g., `experimental_shell_command` and `experimental_run_shell_command`).  It is basically `experimental_run_shell_command` but for the `test` goal instead of the `run` goal.