Fix bug in Go test result exit code interpretation. (#17661)
Gets rid of the confusing, implicit "skipped" state, in favor of checking the necessary field values
directly for None before using them.

Context:

The Go backend performs its own test discovery. It uses None values on the `exit_code` 
and/or `result_metadata` fields to communicate back to the test goal that:

A) Test discovery discovered no tests (exit_code=None, result_metadata=None)
B) Test discovery failed due to compilation error (exit_code=1, result_metadata=None)

Elsewhere, the test goal interprets either field being None as the TestResult being in an
implied "skipped" state.

This lead to a bug (#17657) where a result in state B) was seen as "skipped", and therefore 
did not affect the final pants run exit code, leading to an exit code of 0 in a run that had 
compilation errors, which is a potential CI disaster.

This bug was a result of confusion due to the implicit "skipped" state collapsing the meaning
of None on two different underlying fields into a single concept, that was than misapplied.

This change gets rid of the "skipped" concept entirely. It was misnamed from the start - it does not
represent tests that are skipped but packages that had no tests in them. 

Instead, we test each of the two fields for None-ness before using them, and document when 
each can be None.

PS Note that this issue doesn't occur in the JVM backend, where compilation errors fail the entire run eagerly.
TODO: Should the Go backend do the same thing, and propagate compilation errors immediately?