Makes `tools` optional for `experimental_shell_command` (#17704)
This change looks for the `bash` exit code 127, which indicates that a command could not be found. An error log is raised at that point.

As a side effect, we no longer need to nag the user to set `tools`, which makes `experimental_shell_commands` that depend only on `bash` internal commands a bit more concise.

Note that this would _not_ raise a warning if the error is not one that bash would ignore -- e.g. running a command in backticks, without `set -e` enabled.