More precise running workunit for processes executing remotely (Cherry-pick of #17719) (#17725)
The `run_execute_request` workunit includes the "scheduling" time inside the remote execution server, during which the process might be queued because other workers are busy. But some servers (including `buildgrid`) support sending `Operation` API updates which indicate that a process has moved from queued to actually running.

To avoid including the queueing time in the Pants UI (and blaming processes as "long running" even though they are just queued), this change begins to consume the `ExecuteOperationMetadata` on streaming `Operations` to wait to begin reporting the process as running until either:
1. We see an `ExecutionStage::Unknown` (indicating that the server probably doesn't support sending fine-grained operations)
2. We see an `ExecutionStage::Executing`

The impact is that an overloaded execution server which reports `ExecuteOperationMetadata` will not begin reporting processes as running until they have actually begun executing.