Adds `execution_dependencies` for `experimental_shell_command` (#17743)
As seen in #17345, `experimental_shell_command` has a confused notion of dependencies. There are the "runtime dependencies": those needed to successfully run the command (essentially input data or runnable binaries), and then there are the dependencies that are necessary to use the side-effects generated by the command. `esc` currently treats these identically, which means that the sandboxes for chains of commands get overly large, and their cache invalidations overly fine (due to so many dependencies).

This adds a `SpecialCasedDependencies` field to `experimental_shell_command`, called `execution_dependencies`. `execution_dependencies` are used to populate the sandbox, but are not fetched when `TransitiveTargetsRequests` are made by subsequent targets; `dependencies` is now deprecated and replaced with `output_dependencies`, which are the dependencies necessary for successfully consuming the outputs of the `experimental_shell_command`.

To make it easier to migrate, if `dependencies` are specified _without_ declaring `execution_dependencies`, a deprecation warning will be raised explaining how to upgrade.

Closes #17345