Remove boilerplate for run and test plugins (#17774)
This change provides default rules for `run` and `test` plugins for the `--debug` and `--debug-adapter` flags so that clients don't have to define them unless they opt-in. It also provides us a single exception raised if a feature isn't supported (some of the plugins raised a string "this is a stub", which isn't very user-friendly).

While I was in the neighborhood, I fixed some holdover code from when test plugins registered using `TestFieldSet` as the union.

I don't love adding a `rules()` to a `FieldSet` subclass because some plugin subclasses inherit from multiple field sets. That can be solved with some `super()` magic, but the alternative is to break the `run` API and add a new request type solely for this. Right now this 100% backwards compatible (a feat) so I'm open to comments.