Add export-codegen goal to more backends (#17773)
Several backends create `GenerateSourcesRequest` but do not create the `export-codegen` goal. If these backends are activated alone, that goal will not be registered and thus not work, despite some of their targets supporting it. I imagine this will rarely be relevant, as the backends will mostly used with others that _do_ register the goal.

This PR fixes the ones I can find by `./pants GenerateSourcesRequest --help-advanced` (specifically, comparing that list to the files that mention `export_codegen_goal.rules()`, and adding that as necessary).