Explicitly use "python" ijson backend
ijson provides a few backends with varying levels of compatibility and
performance; which backend is chosen by ijson is system-dependent.
Unfortunately the yajl2 backend does not play nicely with the JSON
output by go list and throws an IncompleteJSONError when parsing.

As a workaround, we can explicitly use the pure Python backend which is
able to parse Go's output correctly.