Restore support for :all: in only_binary/no_binary. (Cherry-pick of #17779) (#17784)
We document that the user can pass `:all:` in `resolves_to_no_binary`/`resolves_to_only_binary`, but #16559 inadvertently broke this. 

That change also had understandably incorrect logic in it: `pkgutil.Requirement`'s `__eq__` method, strangely, does *not* match the PEP 503 name canonicalization that PyPI/pip apply.

This change manually reverts the relevant parts of that change, so that values are passed around as strings again, instead of `PipRequirement` instances (which was overkill anyway, since they are intended to just be project names, not arbitrary requirement strings with version constraints). 

Instead, to solve the issue that #16559 was intended to solve, we canonicalize the option values on input.