go: rename option to --go-test-coverage-packages (#17795)
Rename the `--go-test-coverage-include-patterns` option to `--go-test-coverage-packages` to better align the name with the Go toolchain's name `-coverpkg`. We can rename the option now without deprecation since there has not been a stable release yet of 2.16.

This renaming will eventually allow adding exclusion patterns (prefixed with `!`) to this option which would be a conflicting UX if "include patterns" still remained in the name.