go: detect when `go.sum` needs additional entries filled in (#17811)
As described in https://github.com/pantsbuild/pants/issues/17794, the Go backend will error for inconsistent `go.sum` entries when a module's checksum is different than what is in `go.sum`, but was tolerating _missing_ `go.sum` entries which means that the "source of truth" in `go.sum` was not being updated as it would be if the user were using `go` directly.

Fix this issue by detecting when `go mod download` updates `go.sum` and require the user to update `go.sum` accordingly.

Note: At some point in the future, Pants should manage `go.sum` entries more directly, but that functionality is not needed at the moment to detect and fix this issue.