Detect Pants-generated lockfiles before attempting to validate. (#17833)
Previously we would use the value of the python-only
resolves_generate_lockfiles option to decide whether user
lockfiles were expected to be Pants-generated, and therefore
subject to validation, or not. And we would always validate
tool lockfiles, which were expected to always be Pants-generated.

With this change, the lockfile parsing code raises a new exception
subclass if the lockfile has no "Generated by Pants" header block.
This allows us to differentiate between "this is a Pants-generated
lockfile with invalid metadata", which remains an error, and
"this is not Pants-generated at all", in which case we silently
skip validating metadata.

The only case where this behavior is different from the previous
behavior is if you had a lockfile that did have that header, but
the user wanted Pants not to validate it, which is such a perversion
of the intent of this feature (which is to support lockfiles generated
manually using pex or poetry) that I see no need to worry about it.

This is part of a sequence of planned changes to unify tool and
user lockfiles into a single mechanism.