Unify validation of tool and user lockfile metadata. (#17834)
The only distinction was that user requirements only had to
be a subset of the requirements the lockfile was generated for,
while tool requirements had to be identical. But that distinction
is not significant - a tool will run fine if it is just a subset
of a larger lockfile, so there is no need to be extra strict in
this case.

This not only allows us to reduce complexity, but may also be
useful. For example, it paves the way to using a user lockfile
that includes pytest as the lockfile for running pytest, thus
avoiding having to update pytest versions in two places.