go: support the `go test` profiling options (#17836)
Support the `go test` profiling options which take various profiles of a test run (e.g., CPU, memory, etc.). The profiles are output via the `extra_output` field on `TestResult`. Also includes the ability to export the compiled test binary (which is needed for analysis of some of the profiles).

Closes https://github.com/pantsbuild/pants/issues/16610.