Unify the messages for lockfile validation errors. (#17842)
Now the same error message apply whether the lockfile is
a user or tool lockfile.

This is a step on the road to unifying the two concepts.

Note that these error messages also further clarify the distinction
between how a lockfile was generated vs how it is used:

Today only a lockfile generated from tool requirements can be used
to later satisfy that tool's requirements, so the concept of a "tool
lockfile" is well-defined. But in the future we'll want to, e.g., allow you
to use your python-default lockfile to provide pytest, so you don't
have to configure pytest requirements in two places. So this is a lockfile
generated from user inputs being used to satisfy tool requirements.
In this world the concept of "tool lockfile" is no longer clear.

Also: a tiny docs tweak that didn't justify its own PR.