Use GITHUB_PATH to modify paths between steps.
Previously we used GITHUB_ENV, and interpolated the existing $PATH.

However in the containerized manylinux wheel builds this interpolated
the host $PATH into the container's $PATH. This did not seem to cause
harm in practice on the x86_64 manylinux wheel build, presumably
because the host paths happened to still be suitable in the container.
However this was exposed while working on setting up the aarch64
manylinux wheel build, where the host and container paths are
not at all compatible.