allow each resolve to be exported for multiple python interpreters (#17711)
I want to export a first party resolve multiple times, for multiple interpreters.
The core export code deleted the reldir before toning the export. The python export code passes the resolve export path as reldir, however, so exporting with a different interpreter version forces all old exports to be deleted, even if the interpreter version differs from the old export.

So, this PR includes the interpreter version in reldir. Thus, the core export goal will delete an interpreter specific export only if it is the same as the version that is currently being exported.

This means people will have to clean out old exports themselves, but they will also be able to have multiple exports of the same resolve each with a different interpreter version.