Always stop on `StopIteration` in `rule_runner.run_rule_with_mocks` (#17901)
All `StopIteration` exceptions are, well, stopping. Whether the value is falsy or not shouldn't play a part.

This change was introduced in https://github.com/pantsbuild/pants/pull/8639, I suspect the intent was "if we the generator returns a value then `e.args` will be truthy". However the generator can return Falsy values.

@kaos found this one for me :smile: