Add targets to re-wrap source files in different `SourcesField` types. (#17877)
This is primarily intended to be used with `experimental_shell_command`/`experimental_run_in_sandbox`, but will also work with _any_ target that produces a `SourcesField`/`FileSourceField` that is not applicable for use by dependent rules -- e.g. if your `experimental_run_in_sandbox` produces Python code that you want included as source in a Python distribution. 

The implementation is hacky at the moment, but it mostly solves a problem.

Fixes #15493.