Adds `workdir` field for `experimental_shell_command` and friends (#17928)
This changes the behaviour of `experimental_shell_command` and `experimental_run_in_sandbox`: processes are now run in the buildroot by default, a `workdir` must be specified explicitly to run in any other directory, including the one where the target is defined.

If `workdir` is set, the output files are returned relative to the `workdir`, as before. There is possibly an argument in favour of adding an `output_prefix` field before stabilising.

Fixes #16807.