WSL disclaimer in Getting Started/Prerequisites (#18000)
A short disclaimer was added to inform new Windows users that their projects must live within the Linux virtual machine for Pants to work correctly. 

Feedback welcome! In particular, I want to confirm that using the `wsl` prefix in a Windows shell has the same limitations. I will check myself in a few days if no one knows already.

Related to discussion in #17986.