go: use a directory within sandbox for cgo objects from stdlib packages (#17993)
Usually, the Go backend will choose the package directory as the location for object files resulting from cgo builds. This directory is within the execution sandbox and so this is fine. For standard library packages, however, the package directory is in the `GOROOT` located outside of the execution sandbox, and so this choose fails.

Fix the issue by defining a directory within the execution sandbox as the output directory of cgo builds of standard library packages.

This is part of fixing https://github.com/pantsbuild/pants/issues/17950 and was extracted from https://github.com/pantsbuild/pants/pull/17914.