go: use a Go binary to check for golang assembly files (#17995)
The Go backend when building Cgo-using packages checks whether any Golang format assembly files are present and errors if they are since gcc-format assembly is expected in the Cgo case.

That heuristic check is currently implemented by loading a `Digest` into memory with `DigestContents`. This only works if the files are stored in a digest. For standard library packages, the files live in the `GOROOT` at an absolute path and not in a `Digest`. A naive solution would try to read the files directly from rule code with `open`, but this would obviously fail in remote execution and docker environments.

Fix the issue by using a Go binary to perform the check which can be run locally or remotely as needed.