Make the python dep inference script extensible.
This is so we can implement custom dep parsing logic in plugins, without the performance overhead of running multiple processes per input file.

* Breaks the script up into:
     - A base visitor class for reuse
     - A general-purpose visitor that implements the current, generally-applicable logic
     - A main that executes multiple visitors, as specified by an env var
* Introduces a union for visitor types
* Logic for providing a visitor, and accompanying env var settings, based on union membership

For the existing general-purpose visitor, this meant reading the settings directly from the subsystem, instead of passing them through the request object, which is fine (and could have been implemented that way all along TBH).