Add yamllint as a lint tool for YAML sources (#17821)
This PR adds a backend for running [yamllint](https://pypi.org/project/yamllint/) on YAML source files.

The `yamllint` backend is targetless. It finds all YAML files in the repository matching a configurable regular expression and calls `yamllint` on them. The include and exclude regexps and command line arguments are specified as subsystem options.

The found YAML files are partitioned by configuration files. The plugin searches for files named `.yamllint` by default. This filename is configurable by a subsystem option. A found configuration file is used for the entire subtree of files rooted at the directory that contains it. Closer configuration files take priority. Note that `yamllint` is called with its working directory being the repository root. So, any exclude directives for `yamllint` in a configuration file need to be expressed as full paths from the root.