Ensure we set the region when deploying to S3.
The AWS CLI requires a region, either explicitly on the command line,
via the AWS_DEFAULT_REGION env var, or a default profile.

The publish-tag-to-commit-mapping CI job was failing, apparently
due to lack of a region (as established by SSHing to the CI runner
and reproducing the problem).

It is not yet clear why publishing to S3 works in the wheel publish
jobs, which should have the same issue.