go: rename generated test main output (#18097)
Rename the name of the generated test main for test binaries to remove the leading underscore from the filename. Also, define the name in one place and pass that name through to the generator helper to avoid duplicating the name in several places.

This is necessary because once https://github.com/pantsbuild/pants/pull/17914 lands with Go SDK compilation support, the test goal will begin to run the package analyzer on the generated `testmain.go` and the package analyzer skips files with leading underscores.