go: introduce import maps package attribute (#18098)
Introduce `import_map` on packages which maps to the `ImportMap` key used by `go list`. This is used in vendoring situations (including the standard library) to remap an import path in the source to a canonical import path for finding the actual dependency. (Although building the Go SDK won't be enabled fully until https://github.com/pantsbuild/pants/pull/17914 lands. This is split out from that PR.)