A Cloudflare Worker to handle requests to static.pantsbuild.org.
The worker logs the requests with both Google Analytics 4 and
Universal Analytics, and then redirects to pantsbuild.github.io.

The UA code is the current code we've been using for the last few
years, but which we previously didn't check in to the repo.

Google is phasing out UA, which is why we now add GA4 support.
Once both have overlapped successfully for a while, we'll shut
down the UA side.

I've manually tested this code and verified that requests show up
in both GA4 and UA as expected, and that the right content
is returned.